Got rid of spurious warning about shadowed variables -- they are being translated correctly now (famous last words.)
Added functions to parse an int or a double from a string to
the string library.  I would've used strtoi and strtod but it's
hard to ensure that these don't run off the end of the array.
Also, it appears that strtod and atof are buggy in the version
of Cygwin that I'm using.  (I need these functions for the
GML port that I'm doing.)

Updated cyctest to test the new conversion functions.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@870 581db07f-6e62-4d40-9733-d432c4eebb07