Now everything is automatically wrapped with a "namespace Cyc" unless you specify the -nocyc flag on the command line, or you have an extern "C" declaration.  This will hopefully allow us to avoid name clashes with C.
In addition, cyc_main is now renamed to just main (the Cyc_ gets
added as part of the namespace.)  Also, fixed a bug in the pretty
printer where it was splitting qualified identifiers across lines
(this is okay for Cyclone, but not when we're dumping out C code.)

Next on the list, we should re-think the core and C runtime so
that we don't have silly things like cyc_stdin.

Note that you can still have subtle name clash problems with
extern C declarations -- for instance, doing any of the following
causes all hell to break loose.

int x = 3; // translated to Cyc_x

namespace Foo {
  int x = 2; // translated to Cyc_Foo_x
  extern "C" int x = 4; // translated to x
  extern "C" int Cyc_x = 5; // translated to Cyc_x
}

namespace Bar {
  extern "C" string x = "hello"; // translated to x
}


or if you specify the -nocyc flag then you could run into problems.
I don't know a good way around this -- seems like we'd have to do
a lot more bookkeping for extern C variables to make sure we don't
get a clash.

One annoying property of this change is that the error messages
all mention Cyc::x instead of just x.  We may want to surpress
the leading Cyc within error messages.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@894 581db07f-6e62-4d40-9733-d432c4eebb07