The translation to C was broken if you used goto to leave a try. To fix this, I changed the way the translation figures out how many handlers to pop on break/continue/fallthru/goto/return.
The type-checker was already computing the destination statement for
all these constructs (there isn't one for return obviously).  And it
was trivial to decorate each statement with its try depth.  So the
right number of handlers to pop is just the difference in depths
between source and destination.  This significantly simplified the
environment for translation to C.

Only had to add one kludge to get this to work.  Consider:
  try
    switch(e) {
    case 0: break;
    default: break;
    }
  catch { ... }
  42;

The destination for the breaks was being set to the statement 42;
i.e. across a try block.  But the translation was breaking to an
inserted "pop_handler".  So after my change, the handler would be
double-popped which is big trouble.

The kludge was to rewrite try bodies during type-checking to end with
an extra skip statement, so breaks like the ones above will not cross
handlers.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@973 581db07f-6e62-4d40-9733-d432c4eebb07