* Bug fix with respect to increments/decrements on ? pointers.  So, now the following works:
  string s = "hello\n";
  while (*s) {
    putc(*s++,stdout);
  }

* Some bug fixes w.r.t. attributes.  I'm not exactly sure how GCC
treats them, so I'm guessing in many places.  Note that unlike GCC
we can't allow a declaration to start with __attribute__((...))
because of parsing conflicts.  However, the attributes can follow
just about anything else (e.g., const, extern, int, etc.).

What's tricky is how the attributes get "attached".  In general,
you should try to follow these rules to get things right:

1. for struct declaration, write the attributes (packed or aligned(n))
after the curly brace and before the semi-colon.  For instance:

  struct Foo {short x; short y;} __attribute__((packed));

2. for struct fields, write the attributes after the field name but
before the semi-colon.  For instance:

  struct Foo {short x __attribute__((aligned)); float y;};

3. for function declarations, write the attributes after the return
type.  For instance:

  int __attribute__((dllimport)) foo(int arg) { ... }

4. for function types, write the attributes after the declarator.
For instance:

  int foo(int g(int) __attribute__((stdcall)), int i)

Other positions may work for these things, but they may not.  In
particular, for function declarations, I think you can write the
attributes before the curly brace that starts the body.  GCC doesn't
allow this, but our pretty-printer prints them out in a GCC-friendly
way.  Like GCC, we allow a function declaration to have attributes
that go with the function type (e.g., stdcall) and with the
function declaration (e.g., dllimport).  There's a lot of hackery
to move the type attributes into the type, and keep the function
attributes on the function.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@979 581db07f-6e62-4d40-9733-d432c4eebb07