Mucho hacking, thanks to a lot of travel :-)
* the abstract syntax now uniformly uses "_t" on the typedefs and
I've tried to clean up a few things.  For instance, we have
typ -> type_t, exp -> exp_t, stmt -> stmt_t, etc.

* we can finally name stack-allocated regions.  The region for a
function is given by the unqualified name for that function.  For instance:

  void foo() {
    int x = 0;
    int *`foo y = &x;  // y is a pointer into region `foo
    ...
  }

In addition, labeled blocks (e.g., L:{ ... }) allow you to use the
label name as the region within the block.  For instance:

  L:{ int x = 0;
      int *`L y = &x; // y is a pointer into region `L
      ...
    }

Note that if you duplicate the function name or a label name as an
explicit type parameter then you'll get an error.  This can be
confusing because when you use the same label twice within a function,
instead of getting an error that says "same label used twice" you get
an error that the type variable you're introducing is a duplicate.  Oh
well.

* enum's have reverted to their meaning in C (i.e., symbolic names
for integral constants.)  The old Cyclone enum's have now become
tunion (tagged union), and xenum's have now become xtunion
(extensible tagged union).

* when you declare a tunion such as:

  tunion foo {
    A;
    B(int);
  }

then there are really three types:

* tunion `rgn foo -- a foo value in region rgn.  The default region
  is the heap.

* tunion foo.A -- this is the type of A.  It can be immediately cast
  to tunion `rgn foo for any region since it's a non value-carrying
  constructor.

* tunion foo.B -- this would be the type of B(3).  Note that this is
  not immediately castable to tunion foo.  You have to allocate the
  value first.  That is, tunion foo.B@ is castable to tunion foo.

Note that for xtunions, all constructors must be allocated.
(This turns out to be a pain and not really necessary, so I
might change this if people complain loudly enough.)

So, when adjusting to the change, you need to do the following:

1. replace [x]enum with [x]tunion
2. write "new" in front of any value-carrying constructors.
3. write "&" in front of any value-carrying constructors within a pattern.

In addition, you may have to do a little explicit casting though
I've tried to make this as minimal as possible.  This tends to fail
when you use "let" in combination with a tuple.  In this situation,
there's just not enough context to tell that you wanted to cast up,
and since the fields of the tuple are mutable, we can't (easily) do
the cast at the point where the tuple is used.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@999 581db07f-6e62-4d40-9733-d432c4eebb07