Examined and updated the libc wrappers (stdio, stdlib, time).  I didn't find any real safety problems (but see the last paragraph).
  - Updated time.{h,cyc} to use string rather than char? for
    readability.

  - Optimized all the stdio functions to use underlying_Cstring()
    rather than string_to_Cstring() because the underlying C type is
    const char *.

  - Changed the parameter type of fflush() from FILE@ to FILE*, since
    fflush(NULL) is defined to flush all open output streams.

The Cyclone test suite failed even before my changes, and hello
currently throws an uncaught exception, so I didn't bother writing any
new test cases yet.

Is it good for putenv() to pass the underlying C string to libc as it
currently does?  That sounds dangerous to me (if the string could be
moved or freed by Cyclone later), but it's keeping with the C spirit
of the function: changing the string later will change the environment.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1056 581db07f-6e62-4d40-9733-d432c4eebb07