Fixed a major subtle bug in the type checker that was detailed in the todo file.  The bug was that we allowed, via unification, for type variables to escape their scope.  This had bad interactions with both regions and existentially quantified types, or with '_' types used at the top-level.  This also gets rid of a potential problem with the pretty-printer in that after type-checking, all of the types that are filled in should be well-formed. Details for the fix are below.
In addition, I added support for kind-inference on '_' so that
you can now use '_' just about anywhere.  For instance:

  glist_t<int,_> = ...

The fix for evars was to associate with each evar the set of type
variables that were in scope at the point where the evar is created.
When constraining an evar via unification, we only allow types whose
free type variables are contained in the set associated with the
evar.  It's fairly easy to do this check as we're doing the occurs
check.  As a result, the following code which used to go through:

3:   let x = null;
4:   let init = 0;
5:   while (1) {
6:     region r {
7:       if (!init) {
8:         x = rnew(r)(3); init = 1;
9:       } else {
10:        *x++;
11:      }
12:    }
13:  }

and which is obviously unsound now gives a type error as it should:

  foo.cyc(8:8-8:22): type mismatch: int *%(1)::R  != int *`r

These errors can be confusing because it looks like the region
`r should be able to instantiate the evar %(1).  More information
can be had by passing the -printfullevars flag to the compiler:

  foo.cyc(8:8-8:22): type mismatch: int *%(1<`bug,#0>)::R  != int *`r

here we see that the evar %(1) can only be instantiated in a
context where the region variables `bug and #0 are allowed to
be free.  Since `r isn't in the set, we get the resulting error.

To make this all work, I had to substantially change tcutil.cyc
and make sure that whenever we created an evar, we knew the
right context.  I fear that these changes substantially slowed
down the compiler.  I'll try to do some performance tuning later
on.  There may be a way to do this using a simple counter
instead of sets of type variables, but I couldn't figure out
how to do it without severely restricting unification.  This
is the obvious "right" solution --- I just need to go back
and tune a bit more carefully.

Also, I had to deal with alpha-conversion issues for type variables.
Now type variables are given a unique identifier corresponding
to a scope.  That is, we alpha-vary type variables internally
by assigning them appropriate id's (an integer) during the
well-formedness checks.  The problem this avoids is that we don't want
to confuse two different type variables when doing the checks
on evars.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1092 581db07f-6e62-4d40-9733-d432c4eebb07