[Would someone please look at bin/genfiles/Makefile?  I had to add an explicit .exe on the copy command which installs cyclone.exe into bin/.  The $(EXE) thing wasn't working on Win32.  My change may have broken things under Linux...]
I got rid of the -fwritable-strings requirement.  Now, just as in C,
string literals are treated as "const char" arrays instead of
char arrays.  I also updated all of the libraries and the compiler
to use "const" on strings more appropriately.  For instance, most
of the string library now takes const char arrays as parameters
(unless it needs to update them.)

To expedite the change, note that the typedef "string" no longer
exists.  Rather, we have "string_t" which is constant and
"mstring_t" which is mutable.  In particular, precore.h defines:

string_t = const char ?
stringptr_t = const char ? @
mstring_t = char ?
mstringptr_t = char ? @

So when you write "string_t" now, keep in mind that the string
cannot be updated.

I had to modify the compiler in a number of places (e.g., format
string stuff) to make the changes go through.  I also had to
modify the libraries, test files, etc.  Hence the global changes
(sorry about that.)

One useful trick is that if you have an string_t and need to
get an mstring_t, then use String::dup which copies the string.
I only had to do this in one or two places throughout the system.

An interesting problem that came up is that gcc doesn't allow
us to cast structs.  You can cast pointers to structs but not
structs.  As a result, we can't (easily) represent a string_t
using a struct with const fields.  In fact, right now, the
translation ignores the qualifiers and always uses non-const
fields for ? pointers.  My changes required the insertion of
a few judicious casts when initializing these structs to avoid
warnings.

None of this has been tested under VC++ or Linux.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1098 581db07f-6e62-4d40-9733-d432c4eebb07