Renamed cyclib.a to libcyc.a (following gcc conventions). Do make clean BEFORE you do cvs update, or else old cyclib.a's will stick around.
Beefed up the allocation profiler.  To do allocation
profiling, after make, do make aprof.  This creates
a special version of the libraries, plus a program,
bin/aprof.  Now compile the file you want to profile
with flag -pa.  When you run the program, it creates
a file amon.out with the profiling record.  Run aprof
in the same directory and it will spit out a report.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1172 581db07f-6e62-4d40-9733-d432c4eebb07