(Tries to get a commit in before someone else blows him away...)
Printf, scanf, etc. are no longer baked in, but rather coded as libraries
using the vararg stuff.  Added support for rxprintf so you can region-
allocate the resulting string.  Made a few changes here and there to
the type-checker to get the varargs stuff to work better for these cases.

So, printf, scanf, etc. are no longer keywords and you must include
stdio.h and open it to get printf, etc.  Note that the old code for
type-checking printf and friends is still around because I want to
adapt this so that we can do some static type-checking when the
descriptor string is available.  I'll do this in the near future.

The code behaves the same at least w.r.t. the compiler and the tests
that Trevor coded up in cyctest (including the scanf cases which are
now safe.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1180 581db07f-6e62-4d40-9733-d432c4eebb07