Rewrote main to alternate between up_opt and down_opt until reaching a fixpoint.  This is extremeley dangerous -- in practice we should only iterate a few times, but I like it for now.
Changed Makefile to use -O2, speedup is quite noticeable.

The up-optimizer does better with EM now, that was easy.

The up-optimizer now does some crazy things to set up more for the
down optimizer:  It does the same thing for colors and sizes, I'll
describe it with colors:

If at some point in processing a sequence, it knows that all the
containing text is guarded by some color *and* there are at least two
occurrences of some particular color (say red) that are not nested
inside another color, then it wraps the sequence it has seen so far in
a new red tag.  That way the down optimizer can remove the nested
reds.

On randomly-generated files, the win is pretty small, but it's
guarunteed to be a win.  And on stylized examples, it's awesome:  It
takes
<2>x</2><3><5>z</5></3><2>y</2><2>z</2>
to
<2>x<5>z</5>yz</2>

Three complications definitely worth explaining:

* Splitting an array would be expensive, so Seq now carries an
upper-bound field rather than using the one in the fat-pointer.  NEVER
USE ds.size!  So splitting the array at position i is this abomination
(where dsp and bsp point to the Seq's array and bounds fields)
	doc_t ? before_ds = new {for k < 2: newd};
	before_ds[0] = new Seq(*dsp,i);
	(*dsp)[i] = new Tagged(split_tag,new Seq(before_ds,2));
	*dsp += i;
	*bdp -= i;
Clearly, an unaliased list would have been more convenient.

* It's extremely subtle that the iteration will terminate.  For
  example, if you allow nested occurrences of a color to count, then
  the system will loop on
  <2><3>a<2>b</2></3><4>c<2>d</2></4>f</2>

  Basically, the up will put in a 2 tag that the down will remove.
  Ugh.  By requiring non-nested occurrences, we can't run into this
  problem because the occurrences would have already been eliminated
  by the previous down iteration.  Weird.

* You can't do this for tags like B right now because of the way I did
  PL in the up-optimizer.  This may be fixable, but if you do the
  naive thing, you'll loop forever.

If this is too slow (it didn't seem like it), it's easy to remove.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1231 581db07f-6e62-4d40-9733-d432c4eebb07