I've added cross-compiler support so that we can both compile and distribute .c files for multiple architectures.  Cross-compiling boils down to preprocessing correctly.  Basically, if we are compiling for architecture X, we need to have all of the default macros for that architecture in the environment when we preprocess, and we have to remove the default macros for the architecture we're compiling on.  This is supported as follows:
1) Before building Cyclone, you need to create a customized Makefile.inc.
This is done by doing ./configure in the toplevel directory.  I've removed
the old Makefile.inc from CVS and replaced it with Makefile.inc.in.  If
you need to merge changes you're working on with Makefile.inc, add them
to Makefile.inc.in.

./configure does two things:

  a) It figures out what architecture you're running on and checks to
  see if we know about it.  It uses the autoconf "config.guess"
  script to generate a name like i686-unknown-cygwin1.3.2.  It checks
  to see if a file exists in config/arch that has this name.  This
  file contains the preprocessor default environment.  If this
  file doesn't exist, we need to do more to support the architecture;
  see below.

  b) It creates the custom Makefile.inc by starting from Makefile.inc.in
  and adding two variable definitions, ARCH and CYCDIR.  The former
  is, e.g. i686-unknown-cygwin1.3.2., and the latter is the absolute
  path of the location of the cyclone source tree (this is used
  by other Makefiles in the source tree).

2) All of the genfiles are now stored in bin/genfiles/ARCH, where
ARCH is your architecture name.  Therefore, we will have separate
genfiles for all supported architectures.  Currently I have
Linux and Cygwin.  The old bin/genfiles/{lib,src,Makefile} are still
there in case people have to merge, but these should eventually
be removed.  The actually executable stuff in cyc-lib is NOT
parameterized by architecture, since it should only ever be
created for the architecture you are compiling on.

3) If you want to build for your architecture, do exactly as you
did before.  All of the output is parameterized by the ARCH variable
in Makefile.inc.  If you want to cross-compile, set the Makefile
variable TARGET to be the architecture you want to target.  For example,
you might do

  make cyclone_src TARGET=i686-unknown-linux

and this will compile the compiler using the Linux preprocessor info.
To install these files, you would do

  make update TARGET=i686-unknown-linux

Exact procedures for cross-compiling are in the new toplevel
file INSTALL.

4) The way cross-compiling works is by using a custom preprocessor.
I added a flag -use-cpp to cyclone.exe to support this.  For
cross-compiling we use the preprocessor bin/cyccpp.  This file
expects that its first two arguments will be files of the format
stored in config/arch. The first file is the architecture doing
the compiling (the source architecture), and the second file
is the architecture that we are compiling for (the target
architecture).  The remaining arguments are the ones given to
the standard cpp.  This script will call the standard cpp after
first undefining all macros for the source architecture and
defining those for the target architecture.  This is all built
into the Makefiles in src/, lib/, and lib/xml.

It seems like we should, every time we do a check in, compile for
both Linux and Cygwin.  Other architectures can wait until we do
a public release, and then we can just cross-compile for these.
If something doesn't work, I apologize.  I tried testing everything
on both Windows and Linux, and didn't run into any problems.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1322 581db07f-6e62-4d40-9733-d432c4eebb07