Fixed bug in the configure script.  In addition to defining the the architecture's standard #define's in config/arch/ARCH, it also includes in that file the architecture's name ARCH, but with the - and .'s removed.  This is useful for setting up platform-specific cyclone header files (i.e. for signals, etc.).
Also fixed some things with the interface checker.  In particular, I
made some things extern that were not (like declarations to xtunion exn),
and relaxed the interface checking rules so that structs need not be
extern, to help out with C compatibility.  There is still more work
to do here, though.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1325 581db07f-6e62-4d40-9733-d432c4eebb07