More Makefile clean-up.  In particular, made the tools/*/Makefile files much more similar looking.  Note that they all explicitly name $(CYCDIR)/bin/cyclone.exe -B$(CYCDIR)/bin/cyc-lib -I$(CYCDIR)/include.
Advantage: This allows building the tools before installing the
compiler.  Otherwise we'd need to build the compiler, install the
compiler, build the tools, install the tools.

Disadvantage: These are our simplest Makefiles and it would be nice to
encourage users to copy their form.  Once the compiler is installed,
the -B and -I flags are a bad idea -- you should use the installed
versions.

Personally, I'd like to change this, but it's not a big deal.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1358 581db07f-6e62-4d40-9733-d432c4eebb07