You no longer have to declare structs, enums, unions, tunions, or xtunions before using them in a type definition.  Well, almost. Anyway, the following now works:
struct Foo {
  struct Bar *x;
  int y;
};

struct Bar {
  struct Foo *a;
  int b;
};

Note that I didn't have to start off with a "struct Bar;".  This
should allow more C code to go through unchanged.

There are a few caveats however:  First, if the type is parameterized,
then you do have to declare it before using it.  The reason is that
I need to know what the kinds of the parameters are.  If you try
to do this, then you get a warning message that says you should
declare the struct/tunion/etc. before using it.

The other issue is namespaces.  Right now, the object is declared
for you in the same name space that you're currently in.  This
is the easiest thing to do, but it's not clear to me that it's
the right thing to do.

Took advantage of the new facility to trim out all of the forward
declarations of types in absyn.h (except for the parameterized
ones.)  This had the side-effect of reducing a lot of the genfiles
since the whole compiler includes absyn.h.

Noticed a bug in the control-flow checker for "catch" phrases:
If you leave a "break" off the last case in a catch, then it
re-raises the exception.  Probably a bug that I introduced when
I was trying to allow you to leave break off the last case of
a switch.  I haven't fixed this yet.

Fixed a few more issues with the expanded error messages I checked
in yesterday.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1495 581db07f-6e62-4d40-9733-d432c4eebb07