We now use the flow analysis to eliminate unnecessary null-checks. There is more to do, but we get a lot of the obvious cases.  Of course, complain to me if your Cyclone code starts seg-faulting due to a bug in the analysis. :-)
To do this I added an annotation field to each Exp.  That makes the
abstract syntax significantly larger.  I only need the field on
pointers, so I could add it to certain RawExp variants instead, but
this way I've burned the word once-and-for-all.  Sorry for the
slowdown.

Fixed a soundness bug in the flow-analysis: on ++x and x+=e, I wasn't
invalidating zero/non-zero about x.

Another soundness bug fixed in the flow-analysis: g=e for global g was
making the continuation unreachable.  Don't ask why I screwed that up.

And a very cool "bug" fixed in toc.  The analysis compalain about:

if(field == NULL)
  toc_impos("is_poly_field: bad field %s",*field->name);

for good reason.  Of course, the situation is "impossible", so we
didn't hit the null-dereference in practice.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1502 581db07f-6e62-4d40-9733-d432c4eebb07