Fixed bug in control-flow checker triggered by a union field being a struct.  Fixed it by being extremely lenient about unions -- the analysis assumes they're initialized even if they aren't.  This can't violate memory safety but it does allow reading stack stuff you shouldn't.
Unions are a real pain to deal with unless we enforced the ANSI C rule
that you only read through the last variant you wrote to.  This
prevents some common idioms and is very difficult to enforce in the
presence of aliasing.  So for now, forget it.

compress42.cyc now compiles after adding one initializer because of
the flow analysis being too weak.  In particular, there is data
correlation between successive if-guards.  (James, I added =0 to line
1616.)


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1506 581db07f-6e62-4d40-9733-d432c4eebb07