Three changes to help in porting:
1. 0 can now be used in place of NULL for most things.  The times to
watch out for this are in pattern matching and in printf/scanf-like
functions.  More type information helps the type-checker figure things
out, so if you cast 0 to a pointer type, you'll generally have better
results.  But, for instance, list.cyc compiles if you replace all
occurrences of NULL with 0 so it seems to work pretty well.

2. string_to_Cstring no longer allocates.  Rather, it checks that
the string is 0 terminated and if so, returns a pointer to the
underlying C string.  If not, an exception is raised.

3. &x[i] now works properly.  In particular, it is equivalent to
(x+i).  In fact, that's what it's re-written to.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1550 581db07f-6e62-4d40-9733-d432c4eebb07