Added sizeof_t<t>.  The expression sizeof(t) has type sizeof_t<t> and sizeof_t<t> can be promoted to int (or really any other ingegral type.)  You can, of course, have sizeof_t<`a::A> which is useful for describing the size of an abstract type `a.
Added memcpy and memmove.  These require that you separately pass
in the number of elements you want to copy and the sizeof_t the
elements, unlike C.  The wrappers just do some bounds checking
and then call C's memcpy/memmove so it should be relatively
efficient for large data structures.

Added the ability to cast a "bits-only" ? pointer to a char ?
pointer.  This in turn allows me to give interfaces to
memcmp, memchr, and memset.  Basically, you cast your
bits only pointer to a char pointer, and then call the
appropriate function on it.  Note that memchr isn't all
that satisfactory since you get back a char pointer and
not a pointer to the original type.  Note also that there
are two versions of memchr (memchr and mmemchr) to deal
with the const/non-const issue we have for these sorts
of functions.

The ability to cast bits-only to ? types should help for
other things as well.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1581 581db07f-6e62-4d40-9733-d432c4eebb07