Added better support for malloc and calloc.  You can now malloc arrays, but the type has to be bits-only and you still have to write an explicit:  malloc(sizeof(t)*e) or malloc(e*sizeof(t)).  You can call calloc(sizeof(t),e) as well -- this only requires that t be "zeroable" which includes both * and ? types.
I couldn't figure out how to add GC_calloc and GC_calloc_atomic
in a way that worked well with nogc.  So what I did was add
these to runtime_cyc.c and all they do is call GC_malloc
and GC_malloc_atomic respectively.  Note that I'm counting
on these calls zero'ing memory.  This is the case for the
collector.  It is now also the case for the nogc.c implementation.

If we ever straighten out the makefiles, then we should be
able to add a calloc.c and nogc_calloc.c as appropriate.

I fixed the shootout tests to use thse features but we
can probably minimize the deltas more.  In so doing, I
seem to have broken the nocheck versions of these benchmarks.

But I wanted to get this checked in.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@1750 581db07f-6e62-4d40-9733-d432c4eebb07